requirements.txt
ansi2html==1.9.1
appdirs==1.4.4
asttokens==2.4.1
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.0
frozendict==2.3.10
h11==0.14.0
html5lib==1.1
idna==3.6
importlib-metadata==7.0.0
inflection==0.5.1
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
more-itertools==10.2.0
multitasking==0.0.11
Nasdaq-Data-Link==1.0.4
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
peewee==3.17.0
pexpect==4.9.0
platformdirs==4.1.0
plotly==5.18.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
retrying==1.3.4
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tenacity==8.2.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
watchdog==3.0.0
wcwidth==0.2.12
webencodings==0.5.1
Werkzeug==3.0.1
yfinance==0.2.33
zipp==3.17.0
